+++ src/onegov.feriennet/onegov/feriennet/layout.py
@@ -5,7 +5,9 @@ from onegov.feriennet import security
 from onegov.feriennet.collections import NotificationTemplateCollection
 from onegov.feriennet.collections import VacationActivityCollection
 from onegov.feriennet.const import OWNER_EDITABLE_STATES
-from onegov.org.elements import Link, ConfirmLink, DeleteLink
+from onegov.feriennet.models import VacationActivity
+from onegov.org.new_elements import Link, Confirm, Intercooler, Block
+from onegov.org.new_elements import LinkGroup
 from onegov.org.layout import DefaultLayout as BaseLayout
 from onegov.ticket import TicketCollection
 
@@ -48,8 +50,57 @@ class VacationActivityCollectionLayout(DefaultLayout):
         yield Link(
             text=_("Submit Activity"),
             url=self.request.link(self.model, name='neu'),
-            classes=('new-activity', )
+            attrs={'class': 'new-activity'}
         )
+
+        yield self.offer_again_links
+
+    @property
+    def offer_again_links(self):
+        q = self.app.session().query(VacationActivity)
+        q = q.filter_by(username=self.request.current_username)
+        q = q.filter_by(state='archived')
+        q = q.with_entities(
+            VacationActivity.title,
+            VacationActivity.name,
+        )
+        q = q.order_by(VacationActivity.order)
+
+        activities = tuple(q)
+
+        def link_for_activity(activity):
+            return Link(
+                text=activity.title,
+                url=self.request.class_link(
+                    VacationActivity,
+                    {'name': activity.name},
+                    name="erneut-anbieten"
+                ),
+                traits=(
+                    Confirm(
+                        _(
+                            'Do you really want to provide "${title}" again?',
+                            mapping={'title': activity.title}
+                        ),
+                        _("You will have to request publication again"),
+                        _("Provide Again")
+                    ),
+                    Intercooler(
+                        request_method="POST",
+                        redirect_after=self.request.class_link(
+                            VacationActivity, {'name': activity.name},
+                        )
+                    )
+                )
+            )
+
+        if activities:
+            return LinkGroup(
+                _("Provide activity again"),
+                tuple(link_for_activity(a) for a in activities),
+                right_side=False,
+                classes=('provide-activity-again', )
+            )
 
     @cached_property
     def editbar_links(self):
@@ -164,75 +215,96 @@ class VacationActivityLayout(DefaultLayout):
 
             if self.model.state == 'preview':
                 if period and self.model.has_occasion_in_period(period):
-                    links.append(ConfirmLink(
+                    links.append(Link(
                         text=_("Request Publication"),
                         url=self.request.link(self.model, name='beantragen'),
-                        confirm=_(
-                            "Do you really want to request publication?"
-                        ),
-                        extra_information=_("This cannot be undone."),
-                        classes=('confirm', 'request-publication'),
-                        yes_button_text=_("Request Publication")
+                        attrs={'class': 'request-publication'},
+                        traits=(
+                            Confirm(
+                                _(
+                                    "Do you really want to request "
+                                    "publication?"
+                                ),
+                                _("This cannot be undone."),
+                                _("Request Publication")
+                            ),
+                            Intercooler(
+                                request_method="POST",
+                                redirect_after=self.request.link(self.model)
+                            )
+                        )
                     ))
                 else:
-                    links.append(ConfirmLink(
+                    links.append(Link(
                         text=_("Request Publication"),
-                        url=self.request.link(self.model, name='beantragen'),
-                        confirm=_(
-                            "Please add at least one occasion "
-                            "before requesting publication."
-                        ),
-                        classes=('confirm', 'request-publication'),
+                        url='#',
+                        attrs={'class': 'request-publication'},
+                        traits=(
+                            Block(
+                                _(
+                                    "Please add at least one occasion "
+                                    "before requesting publication."
+                                )
+                            ),
+                        )
                     ))
 
-                links.append(DeleteLink(
+                links.append(Link(
                     text=_("Discard Activity"),
-                    url=self.csrf_protected_url(
-                        self.request.link(self.model)
-                    ),
-                    confirm=_(
-                        'Do you really want to discrd "${title}"?', mapping={
-                            'title': self.model.title
-                        }
-                    ),
-                    extra_information=_("This cannot be undone."),
-                    redirect_after=self.request.class_link(
-                        VacationActivityCollection
-                    ),
-                    yes_button_text=_("Discard Activity"),
-                    classes=('confirm', 'delete-link')
+                    url=self.csrf_protected_url(self.request.link(self.model)),
+                    attrs={'class': 'delete-link'},
+                    traits=(
+                        Confirm(
+                            _(
+                                'Do you really want to discard "${title}"?',
+                                mapping={'title': self.model.title}
+                            ),
+                            _("This cannot be undone."),
+                            _("Discard Activity")
+                        ),
+                        Intercooler(
+                            request_method="DELETE",
+                            redirect_after=self.request.class_link(
+                                VacationActivityCollection
+                            )
+                        )
+                    )
                 ))
 
             links.append(Link(
                 text=_("Edit Activity"),
                 url=self.request.link(self.model, name='bearbeiten'),
-                classes=('edit-link', )
+                attrs={'class': 'edit-link'}
             ))
 
             if not self.request.app.periods:
-                links.append(ConfirmLink(
+                links.append(Link(
                     text=_("New Occasion"),
                     url='#',
-                    confirm=_("Occasions cannot be created yet"),
-                    extra_information=_(
-                        "There are no periods defined yet. At least one "
-                        "period needs to be defined before occasions can "
-                        "be created"
-                    ),
-                    classes=('confirm', 'new-occasion', )
+                    attrs={'class': 'new-occasion'},
+                    traits=(
+                        Block(
+                            _("Occasions cannot be created yet"),
+                            _(
+                                "There are no periods defined yet. At least "
+                                "one period needs to be defined before "
+                                "occasions can be created"
+                            )
+                        )
+                    )
                 ))
             else:
                 links.append(Link(
                     text=_("New Occasion"),
                     url=self.request.link(self.model, 'neue-durchfuehrung'),
-                    classes=('new-occasion', )
+                    attrs={'class': 'new-occasion'}
                 ))
 
             if self.request.is_admin and self.ticket:
                 links.append(Link(
                     text=_("Show Ticket"),
                     url=self.request.link(self.ticket),
-                    classes=('show-ticket', )
+                    attrs={'class': 'show-ticket'}
                 ))
 
             return links
@@ -257,7 +329,7 @@ class PeriodCollectionLayout(DefaultLayout):
             Link(
                 _("New Period"),
                 self.request.link(self.model, 'neu'),
-                classes=('new-period', )
+                attrs={'class': 'new-period'}
             ),
         )
 
@@ -321,7 +393,7 @@ class BillingCollectionLayout(DefaultLayout):
             Link(
                 _("Import Bank Statement"),
                 self.request.link(self.model, 'import'),
-                classes=('import', )
+                attrs={'class': 'import'}
             ),
         )
 
@@ -401,7 +473,7 @@ class NotificationTemplateCollectionLayout(DefaultLayout):
                 Link(
                     _("New Notification Template"),
                     self.request.link(self.model, 'neu'),
-                    classes=('new-notification', )
+                    attrs={'class': 'new-notification'}
                 ),
             )